chore: Separated second notification queue
